refactor(events): replace boilerplate broadcast handlers with generic BroadcastEventHandler<T>
Closes #293
