--- src/Ruvarr/Downloads/Events/DownloadCompletedEventHandler.cs
@@ -1,12 +1,0 @@
-using Ruvarr.Abstractions;
-
-namespace Ruvarr.Downloads.Events;
-
-internal sealed class DownloadCompletedEventHandler(IDomainEventBroadcaster broadcaster) : IDomainEventHandler<DownloadCompletedEvent>
-{
-    public Task Handle(DownloadCompletedEvent @event, CancellationToken cancellationToken)
-    {
-        broadcaster.Publish(@event);
-        return Task.CompletedTask;
-    }
-}
