+++ src/Ruvarr/Downloads/ServiceCollectionExtensions.cs
@@ -15,8 +15,8 @@ internal static class ServiceCollectionExtensions
         services.AddTransient<IRequestHandler<GetDownloadQueueQuery, List<DownloadQueueItemSummary>>, GetDownloadQueueHandler>();
         services.AddTransient<IDomainEventHandler<EpisodeMissingEvent>, EpisodeMissingEventHandler>();
         services.AddTransient<IDomainEventHandler<EpisodeDownloadRequestedEvent>, EpisodeDownloadRequestedEventHandler>();
-        services.AddTransient<IDomainEventHandler<DownloadStartedEvent>, DownloadStartedEventHandler>();
-        services.AddTransient<IDomainEventHandler<DownloadCompletedEvent>, DownloadCompletedEventHandler>();
+        services.AddTransient<IDomainEventHandler<DownloadStartedEvent>, BroadcastEventHandler<DownloadStartedEvent>>();
+        services.AddTransient<IDomainEventHandler<DownloadCompletedEvent>, BroadcastEventHandler<DownloadCompletedEvent>>();
 
         return services;
     }
