--- src/Ruvarr/Programs/Events/ProgramCreatedEventHandler.cs
@@ -1,12 +1,0 @@
-using Ruvarr.Abstractions;
-
-namespace Ruvarr.Programs.Events;
-
-internal sealed class ProgramCreatedEventHandler(IDomainEventBroadcaster broadcaster) : IDomainEventHandler<ProgramCreatedEvent>
-{
-    public Task Handle(ProgramCreatedEvent @event, CancellationToken cancellationToken)
-    {
-        broadcaster.Publish(@event);
-        return Task.CompletedTask;
-    }
-}
