--- src/Ruvarr/Programs/Events/ProgramMatchedTmdbEventHandler.cs
@@ -1,12 +1,0 @@
-using Ruvarr.Abstractions;
-
-namespace Ruvarr.Programs.Events;
-
-internal sealed class ProgramMatchedTmdbEventHandler(IDomainEventBroadcaster broadcaster) : IDomainEventHandler<ProgramMatchedTmdbEvent>
-{
-    public Task Handle(ProgramMatchedTmdbEvent @event, CancellationToken cancellationToken)
-    {
-        broadcaster.Publish(@event);
-        return Task.CompletedTask;
-    }
-}
